fix: separate organization summary hook for fast refresh
